Functional Ellipses Option
Added PagedListRenderOptions.EnableEllipsisNavigation with the default
value of false for backward compatibility.

Modified the HtmlHelper.Ellipses method to use the new render option.
The pager will use the page before/after the last visible page number as
the link target.

Added an example to the "Custom Pager Configurations" section of the
sample site.